+++ packages/std/src/testing/mock.rs
@@ -1,4 +1,7 @@
+use crate::prelude::*;
 use alloc::collections::BTreeMap;
+#[cfg(feature = "cosmwasm_1_3")]
+use alloc::collections::BTreeSet;
 use bech32::{decode, encode, FromBase32, ToBase32, Variant};
 use core::marker::PhantomData;
 #[cfg(feature = "cosmwasm_1_3")]
@@ -8,9 +11,6 @@ use serde::de::DeserializeOwned;
 #[cfg(feature = "stargate")]
 use serde::Serialize;
 use sha2::{Digest, Sha256};
-#[cfg(feature = "cosmwasm_1_3")]
-use std::collections::BTreeSet;
-use std::collections::HashMap;
 
 use crate::addresses::{Addr, CanonicalAddr};
 use crate::binary::Binary;
@@ -240,7 +240,8 @@ impl Api for MockApi {
         )?)
     }
 
-    fn debug(&self, message: &str) {
+    fn debug(&self, #[allow(unused)] message: &str) {
+        #[cfg(feature = "std")]
         println!("{message}");
     }
 }
@@ -700,17 +701,17 @@ impl Default for WasmQuerier {
 #[derive(Clone, Default)]
 pub struct BankQuerier {
     #[allow(dead_code)]
-    /// HashMap<denom, amount>
-    supplies: HashMap<String, Uint128>,
-    /// HashMap<address, coins>
-    balances: HashMap<String, Vec<Coin>>,
+    /// BTreeMap<denom, amount>
+    supplies: BTreeMap<String, Uint128>,
+    /// BTreeMap<address, coins>
+    balances: BTreeMap<String, Vec<Coin>>,
     /// Vec<Metadata>
     denom_metadata: BTreeMap<Vec<u8>, DenomMetadata>,
 }
 
 impl BankQuerier {
     pub fn new(balances: &[(&str, &[Coin])]) -> Self {
-        let balances: HashMap<_, _> = balances
+        let balances: BTreeMap<_, _> = balances
             .iter()
             .map(|(s, c)| (s.to_string(), c.to_vec()))
             .collect();
@@ -740,8 +741,8 @@ impl BankQuerier {
             .collect();
     }
 
-    fn calculate_supplies(balances: &HashMap<String, Vec<Coin>>) -> HashMap<String, Uint128> {
-        let mut supplies = HashMap::new();
+    fn calculate_supplies(balances: &BTreeMap<String, Vec<Coin>>) -> BTreeMap<String, Uint128> {
+        let mut supplies = BTreeMap::new();
 
         let all_coins = balances.iter().flat_map(|(_, coins)| coins);
 
@@ -2186,7 +2187,7 @@ mod tests {
         querier.update_handler(|request| {
             let api = MockApi::default();
             let contract1 = api.addr_make("contract1");
-            let mut storage1 = HashMap::<Binary, Binary>::default();
+            let mut storage1 = BTreeMap::<Binary, Binary>::default();
             storage1.insert(b"the key".into(), b"the value".into());
 
             match request {
@@ -2385,7 +2386,7 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "cosmwasm_1_3")]
+    #[cfg(all(feature = "cosmwasm_1_3", feature = "std"))]
     fn distribution_querier_new_works() {
         let addresses = [
             ("addr0000".to_string(), "addr0001".to_string()),
@@ -2394,7 +2395,7 @@ mod tests {
         let btree_map = BTreeMap::from(addresses.clone());
 
         // should still work with HashMap
-        let hashmap = HashMap::from(addresses.clone());
+        let hashmap = std::collections::HashMap::from(addresses.clone());
         let querier = DistributionQuerier::new(hashmap);
         assert_eq!(querier.withdraw_addresses, btree_map);